Better align initWithValue: and createWithValue: allowed input
For array input init required an exact match for number of values, while
create (implicitly) required that the array have no more elements than
properties and that the primary key (if any) be present. Change init to
match create as it's the backwards-compatible way to make them the same,
and improve the errors thrown by create for the wrong number of values.

Init did not work on single-pass enumerators for array properties due to
trying to pre-validate the values in the array, while create did. Fix
this by using the same validation logic for both. This also makes init
do a deep copy when passed an RLMObject of a different but compatible
type rather than throwing an exception.